Updated NeoVim configuration
Addressed the "TODO" in `init.lua` by adding a utility function to
source a lua file if it exists.

Re-factor some of the lua code in `plugins.lua` and re-order the
contents of `tree-sitter.lua`.

Remove code to enable Vim syntax highlighting and set the sign column to
always be visible.